End-to-end RAG App/requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langchain-community
langchain-core
langchain-groq
langchain-nomic
faiss-cpu
unstructured
python-dotenv
